pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = ""
authors = [
    {name = "Rahul Pandey",email = "pyrahulpndt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "redis (>=6.2.0,<7.0.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "openai (>=1.88.0,<2.0.0)",
    "anthropic (>=0.54.0,<0.55.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "accelerate (>=1.7.0,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "mammoth (>=1.9.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "kombu (>=5.5.4,<6.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "opentelemetry-api (>=1.34.1,<2.0.0)",
    "opentelemetry-sdk (>=1.34.1,<2.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "chardet (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.2.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.1"

